server_dev/requirements.txt
Flask 
Flask-GraphQL
Flask-SQLAlchemy
graphene-sqlalchemy
Flask-Migrate
psycopg2-binary